Moved environment and added stub update class.
