+++ elide-core/src/main/java/com/yahoo/elide/graphql/operations/UpdateOperation.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright 2017, Yahoo Inc.
+ * Licensed under the Apache License, Version 2.0
+ * See LICENSE file in project root for terms.
+ */
+package com.yahoo.elide.graphql.operations;
+
+import com.yahoo.elide.graphql.Environment;
+
+public class UpdateOperation {
+
+    public UpdateOperation(Environment environment) {
+
+    }
+}
